+++ beta/src/components/MDX/Sandpack/DownloadButton.tsx
@@ -92,7 +92,8 @@ ${css}
       onClick={downloadHTML}
       title="Download Sandbox"
       type="button">
-      <IconDownload className="inline mr-1" /> Download
+      <IconDownload className="inline mr-1" />
+      <span className="hidden md:block">Download</span>
     </button>
   );
 };
